port benchmark functions
